ci(cicd.yaml): use action to add
Let's simplify - the script was not correctly pushing the changes (I think). Some googling revealed that there was already an existing action to handle this. To ensure that the commit is properly detedted across jobs, this also splits formatting to its own job, and makes linting, testing, and deployment dependent on it.