fix(cli.py): check for sqlalchemy error
When executing an SQL query, this checks if a ResourceClosedError is thrown. As this is indicative of a successful query that doesnot return (ie CREATE/UPDATE/DELETE/etc.), we can catch it and print a null line instead of throwing the error. If the error is not generated, then we have results, so we print them to the screen.

Closes #12.