feat(cli.py): import csv
If the passed connection string points to a csv, then an in-memory sqlite database is created, and the contents of the csv are dumped to the database. This could cause performance issues with very large databases - both from storing the database in memory, and a speed hit for using pd.DataFrame.to_sql to dump the dataframe. However, performance seems satisfactory on reasonable sized datasets.